+++ src/main/kotlin/dev/aaronhowser/mods/aaron/AaronUtil.kt
@@ -1,6 +1,7 @@
 package dev.aaronhowser.mods.aaron
 
-import net.minecraftforge.common.UsernameCache
+import net.minecraft.nbt.CompoundTag
+import net.neoforged.neoforge.common.UsernameCache
 import java.util.*
 
 object AaronUtil {
@@ -14,6 +15,33 @@ object AaronUtil {
 		}
 
 		return null
+	}
+
+	fun cleanEntityNbt(compoundTag: CompoundTag) {
+		val badTags = listOf(
+			"HurtByTimestamp",
+			"Sitting",
+			"FallFlying",
+			"PortalCooldown",
+			"FallDistance",
+			"InLove",
+			"DeathTime",
+			"UUID",
+			"Age",
+			"ForcedAge",
+			"Motion",
+			"Air",
+			"OnGround",
+			"Rotation",
+			"Pos",
+			"HurtTime",
+			"Owner",
+			"id"
+		)
+
+		for (tag in badTags) {
+			compoundTag.remove(tag)
+		}
 	}
 
 }